Settet the real bank card if radioBankCard is checked. Added if-else logic for which description to take. TT-100

+++ app/src/main/java/hr/foi/techtitans/ttpay/transactions/model_transactions/Card.kt
@@ -3,12 +3,12 @@ package hr.foi.techtitans.ttpay.transactions.model_transactions
 data class Card(
     val cardNumber: String,
     val expirationDate: String,
-    val balance: Int,
+    val balance: Double,
     val cvc: Int
 ) {
 
     // Constructor without conditions
-    constructor() : this("", "", 0, 0)
+    constructor() : this("", "", 0.0, 0)
 
     init {
         require(cardNumber.matches(Regex("\\d{4} \\d{4} \\d{4} \\d{4}"))) {
